Move the completions into a Plugin pile
