+++ src/Development/IDE/Core/Rules.hs
@@ -32,7 +32,6 @@ import Control.Monad
 import Control.Monad.Trans.Class
 import Control.Monad.Trans.Maybe
 import Development.IDE.Core.Compile
-import Development.IDE.Core.Completions
 import Development.IDE.Types.Options
 import Development.IDE.Spans.Calculate
 import Development.IDE.Import.DependencyInformation
@@ -307,20 +306,6 @@ generateCoreRule :: Rules ()
 generateCoreRule =
     define $ \GenerateCore -> generateCore
 
-produceCompletions :: Rules ()
-produceCompletions =
-    define $ \ProduceCompletions file -> do
-        deps <- maybe (TransitiveDependencies [] []) fst <$> useWithStale GetDependencies file
-        tms <- mapMaybe (fmap fst) <$> usesWithStale TypeCheck (transitiveModuleDeps deps)
-        tm <- fmap fst <$> useWithStale TypeCheck file
-        packageState <- fmap (hscEnv . fst) <$> useWithStale GhcSession file
-        case (tm, packageState) of
-            (Just tm', Just packageState') -> do
-                cdata <- liftIO $ cacheDataProducer packageState' (hsc_dflags packageState')
-                                                    (tmrModule tm') (map tmrModule tms)
-                return ([], Just (cdata, tm'))
-            _ -> return ([], Nothing)
-
 generateByteCodeRule :: Rules ()
 generateByteCodeRule =
     define $ \GenerateByteCode file -> do
@@ -378,7 +363,7 @@ mainRule = do
     generateByteCodeRule
     loadGhcSession
     getHieFileRule
-    produceCompletions
+
 
 ------------------------------------------------------------
