+++ src/Development/IDE/LSP/LanguageServer.hs
@@ -30,7 +30,6 @@ import Control.Monad.Extra
 
 import Development.IDE.LSP.HoverDefinition
 import Development.IDE.LSP.CodeAction
-import Development.IDE.LSP.Completions
 import Development.IDE.LSP.Notifications
 import Development.IDE.LSP.Outline
 import Development.IDE.Core.Service
@@ -99,7 +98,6 @@ runLanguageServer options userHandlers getIdeState = do
             setHandlersIgnore <> -- least important
             setHandlersDefinition <> setHandlersHover <>
             setHandlersCodeAction <> setHandlersCodeLens <> -- useful features someone may override
-            setHandlersCompletion <>
             setHandlersOutline <>
             userHandlers <>
             setHandlersNotifications <> -- absolutely critical, join them with user notifications
