+++ src/Development/IDE/Plugin/Completions/Types.hs
@@ -1,5 +1,5 @@
-module Development.IDE.Core.CompletionsTypes (
-  module Development.IDE.Core.CompletionsTypes
+module Development.IDE.Plugin.Completions.Types (
+  module Development.IDE.Plugin.Completions.Types
 ) where
 
 import           Control.DeepSeq
